+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 1. [](#new)
     * Pass phpstan level 1 tests
+    * Updated semver library to v1.5
 
 # v1.10.0-rc.5
 ## 02/03/2020